Changed Leader panel to new format
